requirements.txt
pygithub
pyyaml
colorama
readchar
prettytable

